+++ examples/quick-start.cjs
@@ -33,7 +33,7 @@ async function exampleHandler(header, raw) {
 
 async function main() {
 	sdk = await Kaspeak.create(PRIV_KEY, PREFIX, NETWORK_ID);
-	await sdk.connect();
+	await sdk.connectNode();
 	console.log("Public key:", sdk.publicKey);
 	console.log("Address:", sdk.address);
 
@@ -44,7 +44,7 @@ async function main() {
 	const identifier = SecretIdentifier.random();
 	const tx = await sdk.createTransaction(encoded.length);
 	const opIds = sdk.getOutpointIds(tx);
-	const payload = await sdk.createPayload(opIds, ExampleMessage.messageType, identifier, encoded);
+	const payload = await sdk.createPayload(opIds, ExampleMessage, identifier, encoded);
 	await sdk.sendTransaction(tx, payload);
 }