+++ examples/secret-message.cjs
@@ -20,7 +20,7 @@ class SecretNote extends BaseMessage {
 
 (async () => {
 	const sdk = await Kaspeak.create(6, "TEST");
-	await sdk.connect();
+	await sdk.connectNode();
 
 	sdk.registerMessage(SecretNote, async (h, raw) => {
 		const secret = h.peer.sharedSecret;
@@ -34,6 +34,6 @@ class SecretNote extends BaseMessage {
 	const id = SecretIdentifier.random();
 	const tx = await sdk.createTransaction(body.length);
 	const ids = sdk.getOutpointIds(tx);
-	const plHex = await sdk.createPayload(ids, SecretNote.messageType, id, body);
+	const plHex = await sdk.createPayload(ids, SecretNote, id, body);
 	await sdk.sendTransaction(tx, plHex);
 })();