+++ src/modules/kaspa.ts
@@ -3,9 +3,11 @@ import wasmUrl from "kaspa-wasm/kaspa_bg.wasm?url";
 import * as kaspa from "kaspa-wasm";
 import { DEFAULT_NETWORK_ID } from "../sdk/constants";
 import { bytesToHex, intToHex } from "../crypto";
+import { addressToXPoint } from "../crypto/address";
 import { logger } from "../utils/logger";
 import { EventBus } from "../sdk/event-bus";
 import type { KaspaWasmEvents, Balance, NetworkId, FeeLevel, PaymentOutput } from "../sdk/types";
+import { UtxoProcessorEvent } from "kaspa-wasm";
 
 let kaspaInitialized: boolean = false;
 
@@ -23,14 +25,14 @@ export async function ensureKaspaInitialized(): Promise<void> {
 }
 
 export class KaspaWasm {
-	private _rpc: kaspa.RpcClient | null = null;
-	private _processor: kaspa.UtxoProcessor | null = null;
-	private _context: kaspa.UtxoContext | null = null;
+	private _rpcClient: kaspa.RpcClient | null = null;
+	private _utxoProcessor: kaspa.UtxoProcessor | null = null;
+	private _utxoContext: kaspa.UtxoContext | null = null;
 
 	private _connected = false;
 	private readonly _networkId: kaspa.NetworkId;
 	private _url?: string;
-	private readonly eventBus = new EventBus<KaspaWasmEvents>();
+	private readonly _eventBus = new EventBus<KaspaWasmEvents>();
 
 	private constructor(networkId: NetworkId = DEFAULT_NETWORK_ID) {
 		this._networkId = new kaspa.NetworkId(networkId);
@@ -42,76 +44,79 @@ export class KaspaWasm {
 	}
 
 	public on<E extends keyof KaspaWasmEvents>(e: E, fn: (d: KaspaWasmEvents[E]) => void) {
-		this.eventBus.on(e, fn);
+		this._eventBus.on(e, fn);
 	}
 
 	public off<E extends keyof KaspaWasmEvents>(e: E, fn: (d: KaspaWasmEvents[E]) => void) {
-		this.eventBus.off(e, fn);
+		this._eventBus.off(e, fn);
 	}
 
 	private emit<E extends keyof KaspaWasmEvents>(e: E, d: KaspaWasmEvents[E]) {
-		this.eventBus.emit(e, d);
+		this._eventBus.emit(e, d);
 	}
 
 	public async connect(publicKeyHex: string, url?: string): Promise<void> {
-		if (this._rpc) {
+		if (this._rpcClient) {
 			logger.warn("KaspaWasm is already connected.");
 			return;
 		}
 		if (url) this._url = url;
 		logger.debug(`Connecting to Kaspa network: ${this._networkId} ...`);
-		this._rpc = new kaspa.RpcClient({
+		this._rpcClient = new kaspa.RpcClient({
 			resolver: new kaspa.Resolver(),
 			url: this._url,
 			networkId: this._networkId
 		});
 		this.setUpRpcEventListeners();
-		await this._rpc.connect();
-		this._url = this._rpc.url;
-		kaspa.UtxoProcessor.setUserTransactionMaturityDAA(this._networkId, 10n);
-		this._processor = new kaspa.UtxoProcessor({
+		await this._rpcClient.connect();
+		this._url = this._rpcClient.url;
+		this._utxoProcessor = new kaspa.UtxoProcessor({
 			networkId: this._networkId,
-			rpc: this._rpc
+			rpc: this._rpcClient
 		});
-		this._context = new kaspa.UtxoContext({ processor: this._processor });
-		await this._processor.start();
+		this._utxoContext = new kaspa.UtxoContext({ processor: this._utxoProcessor });
+		await this._utxoProcessor.start();
 		const address = this.getAddressFromPublicKey(publicKeyHex);
-		await this._context.trackAddresses([address]);
+		await this._utxoContext.trackAddresses([address]);
 		this.setUpProcessorEventListeners();
 	}
 
 	public async disconnect(): Promise<void> {
-		if (this._processor) await this._processor.stop();
-		this._processor = null;
-		this._context = null;
-		if (this._rpc) {
-			await this._rpc.disconnect();
+		if (this._utxoProcessor) await this._utxoProcessor.stop();
+		this._utxoProcessor = null;
+		this._utxoContext = null;
+		if (this._rpcClient) {
+			await this._rpcClient.disconnect();
 		}
-		this._rpc = null;
+		this._rpcClient = null;
 		this._connected = false;
 		this._url = undefined;
 	}
 
+	public setTransactionMaturityDAA(maturityDAA: bigint) {
+		kaspa.UtxoProcessor.setUserTransactionMaturityDAA(this._networkId, maturityDAA);
+	}
+
 	private setUpRpcEventListeners(): void {
-		this.rpc.addEventListener("connect", async () => {
+		this.rpcClient.addEventListener("connect", async () => {
 			this._connected = true;
 			logger.debug("Connect");
-			await this.rpc.subscribeBlockAdded();
+			await this.rpcClient.subscribeBlockAdded();
 			this.emit("connect", undefined);
 		});
-		this.rpc.addEventListener("disconnect", async () => {
+		this.rpcClient.addEventListener("disconnect", async () => {
 			this._connected = false;
 			logger.debug("Disconnect");
 			this.emit("disconnect", undefined);
 		});
-		this.rpc.addEventListener("block-added", async (event) => {
+		this.rpcClient.addEventListener("block-added", async (event) => {
 			delete event.data.block.header.parentsByLevel;
 			this.emit("block-added", event.data.block);
 		});
 	}
 
 	private setUpProcessorEventListeners(): void {
-		this.processor.addEventListener("balance", (event) => {
+		this.utxoProcessor.addEventListener("balance", (event) => {
 			if (event.data.balance === undefined) {
 				logger.debug("Balance event received with undefined balance, skipping emit.");
 				return;
@@ -120,23 +125,38 @@ export class KaspaWasm {
 			const utxoCount = event.data.balance.matureUtxoCount + event.data.balance.pendingUtxoCount;
 			this.emit("balance", { balance: balance, utxoCount: utxoCount });
 		});
+
+		this.utxoProcessor.addEventListener("maturity", (event: UtxoProcessorEvent<"maturity">) => {
+			const tx = event.data;
+			const type = tx.type;
+			if (type === "incoming") {
+				logger.debug("Incoming Transaction matured: ", tx.toJSON());
+				this.emit("mature-incoming-tx", { txid: tx.id.toString(), data: tx });
+			}
+		});
+
+		this.utxoProcessor.addEventListener("reorg", (event: UtxoProcessorEvent<"reorg">) => {
+			const tx = event.data;
+			logger.warn("Reorg Transaction event received (consider excluding from non managed lists!): ", tx.toJSON());
+			this.emit("tx-reorg", { txid: tx.id.toString(), data: tx });
+		});
 	}
 
-	private get processor(): kaspa.UtxoProcessor {
-		if (!this._processor) throw new Error("UtxoProcessor is not initialized");
-		return this._processor;
+	get utxoProcessor(): kaspa.UtxoProcessor {
+		if (!this._utxoProcessor) throw new Error("UtxoProcessor is not initialized");
+		return this._utxoProcessor;
 	}
-	private get rpc(): kaspa.RpcClient {
-		if (!this._rpc) throw new Error("Kaspa RPC is not initialized – call connect()");
-		return this._rpc;
+	get rpcClient(): kaspa.RpcClient {
+		if (!this._rpcClient) throw new Error("Kaspa RPC is not initialized – call connect()");
+		return this._rpcClient;
 	}
-	private get context(): kaspa.UtxoContext {
-		if (!this._context) throw new Error("UtxoContext not initialised – call connect()");
-		return this._context;
+	get utxoContext(): kaspa.UtxoContext {
+		if (!this._utxoContext) throw new Error("UtxoContext not initialised – call connect()");
+		return this._utxoContext;
 	}
 
 	public async getUtxosByAddresses(request: kaspa.Address[] | string[]) {
-		return this.rpc.getUtxosByAddresses(request);
+		return this.rpcClient.getUtxosByAddresses(request);
 	}
 
 	public async getBalance(addresses: string | string[]): Promise<Balance> {
@@ -151,13 +171,18 @@ export class KaspaWasm {
 		return key.toAddress(this._networkId).toString();
 	}
 
+	public getXPointFromAddress(addr: string | kaspa.Address): string {
+		const s = typeof addr === "string" ? addr : addr.toString();
+		return addressToXPoint(s);
+	}
+
 	public getPublicKeyFromPrivateKey(privateKey: bigint | number | kaspa.PrivateKey): kaspa.PublicKey {
 		if (privateKey instanceof kaspa.PrivateKey) return privateKey.toPublicKey();
 		return new kaspa.PrivateKey(intToHex(privateKey, 32)).toPublicKey();
 	}
 
 	public async getFeeRate(feeLevel: FeeLevel): Promise<number> {
-		const feeEstimateResonse = await this.rpc.getFeeEstimate();
+		const feeEstimateResonse = await this.rpcClient.getFeeEstimate();
 		const feeEstimate = feeEstimateResonse.estimate;
 		logger.debug(`Estimated fee:`, JSON.stringify(feeEstimate, null, 2));
 		let feeRate: number;
@@ -199,9 +224,10 @@ export class KaspaWasm {
 		}
 		const feeRate = await this.getFeeRate(feeLevel);
 		const dummyPayload = payloadSizeBytes > 0n ? new Uint8Array(Number(payloadSizeBytes)) : undefined;
+		let entries = this.utxoContext.matureLength > 0 ? this.utxoContext : this.utxoContext.getPending();
 		const gen = new kaspa.Generator({
 			networkId: this._networkId.toString(),
-			entries: this.context,
+			entries,
 			changeAddress: myAddress,
 			outputs: transactionOutputs,
 			payload: dummyPayload,
@@ -219,7 +245,7 @@ export class KaspaWasm {
 			tx.payload = payload;
 		}
 		const signed = kaspa.signTransaction(tx, [pk], true);
-		const { transactionId } = await this.rpc.submitTransaction({ transaction: signed });
+		const { transactionId } = await this.rpcClient.submitTransaction({ transaction: signed });
 		return transactionId;
 	}
 
@@ -235,7 +261,15 @@ export class KaspaWasm {
 		return this._url;
 	}
 
+	get kaspa(): typeof import("kaspa-wasm") {
+		return kaspa;
+	}
+
+	get eventBus() {
+		return this._eventBus;
+	}
+
 	public async getServerInfo() {
-		return this.rpc.getServerInfo();
+		return this.rpcClient.getServerInfo();
 	}
 }