+++ docs/ru/05-meta/changelog.md
@@ -1,5 +1,19 @@
 # История изменений
 
+## 1.0.0 — 2025-08-25
+* **BREAKING:** Протокол обновлен до версии 2. Добавлен 1 дополнительный байт для `SignatureType`
+* **BREAKING:** `sdk.createPayload` принимает вторым аргументом не `messageType`, а сам класс сообщения. 
+* **BREAKING:** `sdk.connect` устарело. Используйте `sdk.connectNode`
+* **BREAKING:** Подписки на события `connect`/`disconnect` устарели. Используйте `node-connect`/`node-disconnect` соответственно.
+* Используемые `kaspa.RpcClient`, `kaspa.UtxoProcessor`, `kaspa.UtxoContext` а так же сам проиницализированный модуль `kaspa-wasm` теперь доступен прямо через геттеры экземпляра `sdk` для более тонкого взаимодействия с сетью.
+* Добавлены подписки на события `indexer-connect`, `indexer-disconnect`, `mature-incoming-tx`, `tx-reorg`.
+* Добавлены методы для взаимодействия с индексатором: `sdk.connectIndexer`, `sdk.indexerSend`, `sdk.indexerRequest`
+* Добавлен метод `sdk.getXPointFromAddress`
+* Добавлен метод `sdk.setTransactionMaturityDAA`
+* `sdk.createTransaction` теперь принимает дополнительный необязательный параметр `recipients`, реализующий возможность отправки данных вместе с переводом средств.
+* В `MessageHeader` добавлены дополнительные поля `live`, `signatureType`, `isPayment`, `txOutputs` и опциональный `requestId`.
+* Мелкие исправления и улучшения кодовой базы.
+
 ## 0.1.0 — 2025-07-14
 * **BREAKING:** Исправлена ошибка, не позволяющая отправить сообщение в сети *mainnet*.
 * **BREAKING:** Подписка на событие `KaspeakMessageReceived` переименована в `message`.